common/libjob: Add private job_util function
Move util_jobkey() from job-manager to libjob, so future
convenience functions can be shared between modules.
